requirements.txt
requests
pandas
bs4
python-amazon-sp-api
python-dotenv